+++ backend-2/src/main/java/io/ggogit/ggogit/domain/book/entity/Book.java
@@ -67,7 +67,7 @@ public class Book {
 
     @NotNull
     @Column(name = "TOTAL_PAGE", nullable = false)
-    private Long totalPage;
+    private Integer totalPage;
 
     @Size(max = 255)
     @Column(name = "IMAGE_FILE")
